+++ CHANGELOG.md
@@ -9,10 +9,12 @@ and this project adheres to [Semantic Versioning][semver].
 
 ### Added
 
+- New function taf roles list that lists all roles in an authentication repository ([421])
 - Reimplemented get_file_details function to not rely on old securesystemslib functions ([420])
 - Clone target repositories to temp ([412, 418])
 - Add architecture overview documentation ([405])
 
+[421]: https://github.com/openlawlibrary/taf/pull/421
 [420]: https://github.com/openlawlibrary/taf/pull/420
 [412]: https://github.com/openlawlibrary/taf/pull/412
 [405]: https://github.com/openlawlibrary/taf/pull/405